requirements.txt
boto3==1.26.55
botocore==1.29.55
jmespath==1.0.1
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
